refactor: simplify the EnumMap into a ValueMap
The new idea is that enumerations and bitstrings have IRI-mappings in the logical data schema of the InteractionAffordance. BACnet protocol binding receives a simple logical value as an integer, string or boolean. This can be mapped to an integer for the protocol value. Depending on the BACnet type the mapping can be optional or mandatory.
The same approach is also used for bacnet:CommandPriority. Previously it was an enumeration of IRIs, now it is just an enumeration of integers. There is marginal value in modelling the priority levels as IRIs.
EnumMap is also renamed to a ValueMap, enabling it to be used also for non-enum cases, such as a simple integer mapping or BitStrings.